% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus.R
\name{download_one_file}
\alias{download_one_file}
\title{download_one_file}
\usage{
download_one_file(user_name, date_in, variables_in, save_path = "data")
}
\arguments{
\item{user_name}{The user name for the service}

\item{date_in}{The date associated with the data}

\item{variables_in}{String vector of values to retrieve}

\item{save_path}{Path where the data should be saved. The default is a folder called "data"}
}
\value{
The function will return the name of the downloaded file, as a string
}
\description{
Make a single request to Copernicus.
}
\details{

}
