% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus.R
\name{extract_one_file}
\alias{extract_one_file}
\title{extract_one_file}
\usage{
extract_one_file(file_in, delete_source = TRUE)
}
\arguments{
\item{file_in}{The name of an NC file from which to extract}

\item{delete_source}{Boolean indicating whether to delete the file once the extract has completed}
}
\value{
A tibble (data frame) of the contents
}
\description{
extract_one_file
}
\details{
This does little apart from converting the contents of an NC file into a data frame. Note that this object will likely be very large, on the order of more than one million rows.

If a column called 'time' exists in the data frame, it will be renamed 'cds_time'. This column name is expected in the function \code{add_date_time()}.
}
