% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{summarize_day}
\alias{summarize_day}
\title{summarize_day}
\usage{
summarize_day(tbl_in)
}
\arguments{
\item{tbl_in}{Data frame of climate data to summarize}
}
\value{
A data frame
}
\description{
summarize_day
}
