% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{trim_grid}
\alias{trim_grid}
\title{trim_grid}
\usage{
trim_grid(tbl_in, lat_gran = 0.05, lon_gran = 0.05)
}
\arguments{
\item{tbl_in}{Data frame of climate data}

\item{lat_gran}{Granularity of latitude}

\item{lon_gran}{Granularity of longitude}
}
\description{
trim_grid
}
\details{

}
